work to do with backend

+++ backend/controllers/incomeControllers.ts
@@ -1,13 +1,16 @@
 import { Finanse } from "../models/finanseModel";
-import { Request, Response } from "express";
+import { MyRequest, MyResponse } from "../types/Requests";
+
 import mongoose from "mongoose";
 import { Category } from "../models/categoryModel";
 import { categoryExists } from "./categoryController";
 
 //get all incomes
-export async function showAllIncomes(req: Request, res: Response) {
+export async function showAllIncomes(req: MyRequest, res: MyResponse) {
+  const {user} = req;
+
   //show all incomes and sort them by their date
-  const incomes = await Finanse.find({ finanse: "income" }).sort({
+  const incomes = await Finanse.find({user_id:user, finanse: "income" }).sort({
     createdAt: -1,
   });
 
@@ -15,23 +18,27 @@ export async function showAllIncomes(req: Request, res: Response) {
 }
 
 //get a single income
-export async function showIncome(req: Request, res: Response) {
+export async function showIncome(req: MyRequest, res: MyResponse) {
+
   //get id from req.params (from the path)
   const { id } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ error: "No such income" });
+    return res.status(404).json({ message: "No such income" });
   }
 
   const income = await Finanse.findById(id);
 
   income
     ? res.status(200).json(income)
-    : res.status(400).json({ error: "No such income" });
+    : res.status(400).json({ message: "No such income" });
 }
 
 //create income
-export async function createIncome(req: Request, res: Response) {
+export async function createIncome(req: MyRequest, res: MyResponse) {
+
+  const {user} = req;
+
   //get attributes send in json
   console.log("Create income");
   const { title, amount } = req.body;
@@ -44,13 +51,13 @@ export async function createIncome(req: Request, res: Response) {
   if (emptyFields.length > 0) {
     return res
       .status(400)
-      .json({ error: "Please fill in all the fields", emptyFields });
+      .json({ message: "Please fill in all the fields", emptyFields });
   }
   let category: string = req.body.category;
   category = category && category.toLowerCase();
 
   const categoryExists = await Category.find({category});
-  if (!categoryExists) return res.status(400).json({error: "Category not found"});
+  if (!categoryExists) return res.status(400).json({message: "Category not found"});
   //add to dB
   try {
     const income = await Finanse.create({
@@ -58,55 +65,56 @@ export async function createIncome(req: Request, res: Response) {
       amount,
       finanse: "income",
       category: category ? category : "other",
+      user_id:user,
     });
     res.status(200).json(income);
   } catch (error: any) {
-    res.status(400).json({ error: error.message });
+    res.status(400).json({ message: error.message });
   }
 }
 
 //delete income
-export async function deleteIncome(req: Request, res: Response) {
+export async function deleteIncome(req: MyRequest, res: MyResponse) {
   const { id } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ error: "No such income" });
+    return res.status(404).json({ message: "No such income" });
   }
   const income = await Finanse.findOneAndDelete({ _id: id });
 
   income
-    ? res.status(200).json({ msg: "Income deleted succesfully" })
-    : res.status(404).json({ error: "No such income" });
+    ? res.status(200).json({ message: "Income deleted succesfully" })
+    : res.status(404).json({ message: "No such income" });
 }
 
 //update income
-export async function updateIncome(req: Request, res: Response) {
+export async function updateIncome(req: MyRequest, res: MyResponse) {
   const { id } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ error: "No such income" });
+    return res.status(404).json({ message: "No such income" });
   }
 
   let category: string = req.body.category;
   category = category && category.toLowerCase();
 
-  if (category && !categoryExists(category)) return res.status(400).json({error: "Category not found"});
+  if (category && !categoryExists(category)) return res.status(400).json({message: "Category not found"});
 
   const income = await Finanse.findByIdAndUpdate(id, { ...req.params });
 
   income
     ? res.status(404).json(income)
-    : res.status(404).json({ error: "Failed to update" });
+    : res.status(404).json({ message: "Failed to update" });
 }
 
 //delete all incomes
-export async function deleteAllIncomes(req: Request, res: Response) {
+export async function deleteAllIncomes(req: MyRequest, res: MyResponse) {
   const mongoUri: string = process.env.MONGO_URI as string;
-
+  const {user} = req;
   try {
-    await Finanse.deleteMany({ finanse: "income" });
+    await Finanse.deleteMany({user_id:user, finanse: "income" });
     console.log("incomes deleted succesfully");
-    res.status(400).json({ msg: "incomes deleted succesfully" });
+    res.status(400).json({ message: "incomes deleted succesfully" });
   } catch (error: any) {
     console.log("Failed to delete all incomes");
     throw Error(error.message);
